v0.0.8
* Reimplemented internal ids system.
* Improved sameness detection of Player and Team names.
* Support for Battlefy (LowInk), including additional Player information that it gives.
* Fixed a database editor crash.
* Added a change log.

+++ SplatTagDatabase/DbPlayer.cs
@@ -7,6 +7,6 @@ namespace SplatTagDatabase
   {
     public uint id;
     public string[] names;
-    public uint[] teams;
+    public long[] teams;
   }
 }
